+++ PCIT_core/internal/llvm_interface/src/IRBuilder.cpp
@@ -156,16 +156,39 @@ namespace pcit::llvmint{
 	//////////////////////////////////////////////////////////////////////
 	// type conversion
 
-	auto IRBuilder::createTrunc(const Value& value, const Type& dst_type, evo::CStrProxy name) -> Value {
+	// doesn't use NUW or NSW as for some reason it triggers an assert inside of LLVM
+	auto IRBuilder::createTrunc(const Value& value, const Type& dst_type, evo::CStrProxy name)
+	-> Value {
 		return Value(this->builder->CreateTrunc(value.native(), dst_type.native(), name.c_str()));
+	}
+	auto IRBuilder::createFPTrunc(const Value& value, const Type& dst_type, evo::CStrProxy name)
+	-> Value {
+		return Value(this->builder->CreateFPTrunc(value.native(), dst_type.native(), name.c_str()));
 	}
 	
 	auto IRBuilder::createZExt(const Value& value, const Type& dst_type, evo::CStrProxy name) -> Value {
 		return Value(this->builder->CreateZExt(value.native(), dst_type.native(), name.c_str()));
 	}
-	
 	auto IRBuilder::createSExt(const Value& value, const Type& dst_type, evo::CStrProxy name) -> Value {
 		return Value(this->builder->CreateSExt(value.native(), dst_type.native(), name.c_str()));
+	}
+	auto IRBuilder::createFPExt(const Value& value, const Type& dst_type, evo::CStrProxy name)
+	-> Value {
+		return Value(this->builder->CreateFPExt(value.native(), dst_type.native(), name.c_str()));
+	}
+
+
+	auto IRBuilder::createSIToFP(const Value& value, const Type& dst_type, evo::CStrProxy name) -> Value {
+		return Value(this->builder->CreateSIToFP(value.native(), dst_type.native(), name.c_str()));
+	}
+	auto IRBuilder::createUIToFP(const Value& value, const Type& dst_type, evo::CStrProxy name) -> Value {
+		return Value(this->builder->CreateUIToFP(value.native(), dst_type.native(), name.c_str()));
+	}
+	auto IRBuilder::createFPToSI(const Value& value, const Type& dst_type, evo::CStrProxy name) -> Value {
+		return Value(this->builder->CreateFPToSI(value.native(), dst_type.native(), name.c_str()));
+	}
+	auto IRBuilder::createFPToUI(const Value& value, const Type& dst_type, evo::CStrProxy name) -> Value {
+		return Value(this->builder->CreateFPToUI(value.native(), dst_type.native(), name.c_str()));
 	}